+++ src/main/java/com/ssh/service/PersonService.java
@@ -6,6 +6,6 @@ import com.ssh.entity.Person;
 
 public interface PersonService {
 	
-	int savePerson(Person person);
+	Long savePerson(Person person);
 
 }
